Add packages, move classes to Duke folder
Create Duke folder within java folder and move class files to Duke
folder. Import classess as packages in Duke.java.

Sepearte Task subclasses into their respective files
(e.g todo-> todo.java).

Minor grammar fix in printing total current task(s) as singular where
applicable.
