+++ src/main/java/Duke.java
@@ -1,3 +1,9 @@
+import Duke.DukeException;
+import Duke.Ui;
+import Duke.Parser;
+import Duke.Storage;
+import Duke.Task;
+import Duke.TaskList;
 import java.io.IOException;
 
 public class Duke {
@@ -43,7 +49,6 @@ public class Duke {
                         Task task = list.markDone(order);
                         data.markDoneData(order, task.getParsedData());
                         ui.showDoneTask(task);
-                        ui.showTotalTasks(list.getSize());
 
                     } else if (mode == Parser.Mode.DELETE) {
                         // Delete Task
