+++ src/main/java/Duke/DukeException.java
@@ -1,4 +1,6 @@
-class DukeException extends Exception{
+package Duke;
+
+public class DukeException extends Exception{
     public DukeException(String message){
         super(message);
     }
