+++ src/main/java/Duke/Parser.java
@@ -1,4 +1,6 @@
-class Parser {
+package Duke;
+
+public class Parser {
 
     public enum Mode{
         LIST, BYE, DONE, DELETE, TODO, DEADLINE, EVENT;
