+++ src/main/java/Duke/Storage.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
@@ -5,7 +7,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
-class Storage {
+public class Storage {
 
     // Read file
     private File f;
